+++ lib/src/inspect_io.dart
@@ -91,5 +91,6 @@ Future<TlsCertificateInfo> inspectTls(
         .join(':'),
     trusted: trusted,
     pem: certificate.pem,
+    der: certificate.der,
   );
 }